Refactor stitch view rendering
